IpServiceScanner: Implement recommendation from ETSI TS 102 818 about Caching of logo files
Rationale: Previous implementation never downloaded a new logo again even if it may have changed on the server.
This implementation tries to still minimize the amount of actually downloaded data, but if server does not support to
provide last-modified or HTTP code 304 NOT_MODIFIED, then files may be downloaded unnecessarily.
To provide the user with latest logos is more important than trying to save data traffic.